+++ junit4/src/test/kotlin/fr/xgouchet/elmyr/junit4/ForgeStatementTest.kt
@@ -1,7 +1,11 @@
 package fr.xgouchet.elmyr.junit4
 
+import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.doThrow
 import com.nhaarman.mockitokotlin2.whenever
+import fr.xgouchet.elmyr.annotation.StringForgery
+import fr.xgouchet.elmyr.junit4.dummy.BarFactory
+import fr.xgouchet.elmyr.junit4.dummy.FooFactory
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.AssumptionViolatedException
 import org.junit.Before
@@ -16,23 +20,36 @@ import org.mockito.quality.Strictness
 
 internal class ForgeStatementTest {
 
-    @Rule @JvmField val forge = ForgeRule()
-    @Rule @JvmField val mockito = MockitoJUnit.rule().strictness(Strictness.LENIENT)
-    @Rule @JvmField val systemErrRule = SystemErrRule().enableLog()
+    @Rule
+    @JvmField
+    val forge = ForgeRule().withFactory(FooFactory()).withFactory(BarFactory())
+    @Rule
+    @JvmField
+    val mockito = MockitoJUnit.rule().strictness(Strictness.LENIENT)
+    @Rule
+    @JvmField
+    val systemErrRule = SystemErrRule().enableLog()
 
     internal lateinit var testedStatement: ForgeStatement
 
-    @Mock lateinit var mockBaseStatement: Statement
-    @Mock lateinit var mockMethod: FrameworkMethod
-    @Mock lateinit var mockTarget: Any
+    @Mock
+    lateinit var mockBaseStatement: Statement
+    @Mock
+    lateinit var mockMethod: FrameworkMethod
+    @Mock
+    lateinit var mockTarget: Any
+
+    @StringForgery
+    lateinit var fakeMethodName: String
 
     @Before
     fun setUp() {
+        whenever(mockMethod.name) doReturn fakeMethodName
         testedStatement = ForgeStatement(
-                mockBaseStatement,
-                mockMethod,
-                mockTarget,
-                forge
+            mockBaseStatement,
+            mockMethod,
+            mockTarget,
+            forge
         )
     }
 
@@ -50,13 +67,50 @@ internal class ForgeStatementTest {
         assertThat(e).isNotNull()
 
         assertThat(systemErrRule.log)
-                .isEqualTo(
-                        "<${mockTarget.javaClass.simpleName}.${mockMethod.name}()> failed " +
-                                "with Forge seed 0x${forge.seed.toString(16)}L\n" +
-                                "Add this seed in the ForgeRule in your test class :\n" +
-                                "\n" +
-                                "\tForgeRule forge = new ForgeRule(0x${forge.seed.toString(16)}L);\n" +
-                                "\n")
+            .isEqualTo(
+                "<${mockTarget.javaClass.simpleName}.$fakeMethodName()> failed " +
+                    "with Forge seed 0x${forge.seed.toString(16)}L\n" +
+                    "Add this seed in the ForgeRule in your test class :\n" +
+                    "\n" +
+                    "\tForgeRule forge = new ForgeRule(0x${forge.seed.toString(16)}L);\n" +
+                    "\n"
+            )
+    }
+
+    @Test
+    fun `prints seed and injected fields when failure happens`() {
+        val target = KotlinAnnotationTest()
+        testedStatement = ForgeStatement(
+            mockBaseStatement,
+            mockMethod,
+            target,
+            forge
+        )
+        whenever(mockBaseStatement.evaluate()) doThrow AssertionError()
+
+        val e: AssertionError? = try {
+            testedStatement.evaluate()
+            null
+        } catch (e: AssertionError) {
+            e
+        }
+
+        assertThat(e).isNotNull()
+
+        assertThat(systemErrRule.log)
+            .isEqualTo(
+                "<KotlinAnnotationTest.$fakeMethodName()> failed " +
+                    "with Forge seed 0x${forge.seed.toString(16)}L and:\n" +
+                    "\t- Field KotlinAnnotationTest::fakeFoo = ${target.fakeFoo}\n" +
+                    "\t- Field KotlinAnnotationTest::fakeFooList = ${target.fakeFooList}\n" +
+                    "\t- Field KotlinAnnotationTest::fakeFooMap = ${target.fakeFooMap}\n" +
+                    "\t- Field KotlinAnnotationTest::fakeFooSet = ${target.fakeFooSet}\n" +
+                    "\t- Field KotlinAnnotationTest::fakeMonth = ${target.fakeMonth}\n\n" +
+                    "Add this seed in the ForgeRule in your test class :\n" +
+                    "\n" +
+                    "\tForgeRule forge = new ForgeRule(0x${forge.seed.toString(16)}L);\n" +
+                    "\n"
+            )
     }
 
     @Test
@@ -73,6 +127,6 @@ internal class ForgeStatementTest {
         assertThat(e).isNotNull()
 
         assertThat(systemErrRule.log)
-                .isEmpty()
+            .isEmpty()
     }
 }